feat: add Windows support
Adds the Windows implementation of Testudo.Native. Code has been re-architected slightly to support Windows concepts and new features.

Have also changed from CMake to MSVC because MinGW doesn't have the necessary Windows APIs available to compile for Windows.

BREAKING-CHANGE: Linux support is currently broken due to the change to MSVC and the need to refactor to accommodate the changes made in this commit.